+++ harness/chaos/model_security_governance_time_window.go
@@ -9,9 +9,9 @@
 package chaos
 
 type SecurityGovernanceTimeWindow struct {
-	Duration string `json:"duration,omitempty"`
-	EndTime int32 `json:"endTime,omitempty"`
+	Duration   string                        `json:"duration,omitempty"`
+	EndTime    int64                         `json:"endTime,omitempty"`
 	Recurrence *SecurityGovernanceRecurrence `json:"recurrence,omitempty"`
-	StartTime int32 `json:"startTime"`
-	TimeZone string `json:"timeZone"`
+	StartTime  int64                         `json:"startTime"`
+	TimeZone   string                        `json:"timeZone"`
 }
